Fix blending not working through multiple parents
